+++ RELEASE-NOTES.md
@@ -31,6 +31,7 @@ Initial version with:
 ## [0.7.0](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.6.0...v0.7.0) (2022-07-04)
 * Campaigns > Get campaign
 * Campaigns > Get campaign stops
+* Campaigns > Get campaign npais
 
 ### Upgrade Steps
 * [NO ACTION REQUIRED]
@@ -41,6 +42,7 @@ Initial version with:
 ### New Features
 * Campaigns > Get campaign
 * Campaigns > Get campaign stops
+* Campaigns > Get campaign npais
 
 ### Bug Fixes